Update Skip Types to be generic for any type of component
