+++ packages/SkipTypes/readme.md
@@ -22,7 +22,7 @@ Types used for making requests to the Skip API:
 - `SkipAPIRequestAPIKey` - API key configuration for LLM vendors
 - `SkipAPIRunScriptRequest` - Request to run an existing script
 - `SkipRequestPhase` - Defines the different phases of Skip requests:
-  - `initial_request` - Starting a new conversation or after report creation
+  - `initial_request` - Starting a new conversation or after component creation
   - `clarify_question_response` - Responding to Skip's clarifying questions
   - `data_gathering_response` - Returning requested data to Skip
   - `data_gathering_failure` - Reporting data gathering errors
@@ -51,20 +51,20 @@ Types used for handling responses from the Skip API:
   - Includes id, name, userId, messages array
   - Optional: description, artifacts
 
-### Report Types
+### Component Types
 
-Types for Skip-generated reports and visualizations:
+Types for Skip-generated components and visualizations:
 
-- `SkipHTMLReport` - HTML-based report generated by Skip
-- `SkipHTMLReportCallbacks` - Interface for callbacks from HTML reports:
+- `SkipComponent` - Component generated by Skip
+- `SkipComponentCallbacks` - Interface for callbacks from components:
   - `RefreshData()` - Refresh the data context
   - `OpenEntityRecord(entityName, key)` - Open a specific record
   - `UpdateUserState(userState)` - Update user-specific state
   - `NotifyEvent(eventName, eventData)` - Send custom events
-- `SkipHTMLReportInitFunction` - Initialization function for HTML reports
-- `SkipHTMLReportObject` - Interface exposed by HTML reports
-- `SkipHTMLReportOption` - Alternative HTML report options
-- `SimpleDataContext` - Simplified data context for HTML reports
+- `SkipComponentInitFunction` - Initialization function for components
+- `SkipComponentObject` - Interface exposed by components
+- `SkipComponentOption` - Alternative component options
+- `SimpleDataContext` - Simplified data context for components
 
 ### Learning Cycle Types
 
@@ -96,7 +96,7 @@ Types for Skip-generated reports and visualizations:
 
 - `SkipAPIAnalysisCompleteResponse` - Response for completed analysis
   - Includes dataContext, resultType, executionResults
-  - Optional: htmlReport, drillDown, scriptText, newDataItems
+  - Optional: component, drillDown, scriptText, newDataItems
 - `SkipAPIClarifyingQuestionResponse` - Response when Skip needs clarification
 - `SkipAPIDataRequestResponse` - Response when Skip needs more data
 - `SkipAPIChatWithRecordResponse` - Response for chat-with-record feature
@@ -172,8 +172,8 @@ function handleSkipResponse(result: MJAPISkipResult) {
         // Handle plot data from executionResults.plotData
         renderPlot(response.executionResults.plotData);
       } else if (response.resultType === "html") {
-        // Handle custom HTML report
-        renderHTML(response.htmlReport);
+        // Handle custom component
+        renderComponent(response.component);
       }
     } else if (result.ResponsePhase === SkipResponsePhase.clarifying_question) {
       // Handle clarifying questions
@@ -190,22 +190,22 @@ function handleSkipResponse(result: MJAPISkipResult) {
 }
 ```
 
-### Working with Skip HTML Reports and Callbacks
+### Working with Skip Components and Callbacks
 
 ```typescript
 import { 
-  SkipHTMLReportInitFunction, 
+  SkipComponentInitFunction, 
   SimpleDataContext, 
-  SkipHTMLReportCallbacks 
+  SkipComponentCallbacks 
 } from '@memberjunction/skip-types';
 
 // Init function that would be called by the container application
-const initReport: SkipHTMLReportInitFunction = (
+const initComponent: SkipComponentInitFunction = (
   data: SimpleDataContext, 
   userState?: any, 
-  callbacks?: SkipHTMLReportCallbacks
+  callbacks?: SkipComponentCallbacks
 ) => {
-  // Initialize the report with the data
+  // Initialize the component with the data
   renderChart(data.data_item_1);
   
   // Set up event handlers that use callbacks
@@ -223,7 +223,7 @@ const initReport: SkipHTMLReportInitFunction = (
 };
 
 // Register the init function globally so it can be called from the container
-window.initSkipReport = initReport;
+window.initSkipComponent = initComponent;
 ```
 
 ### Working with Learning Cycles
